CODETOOLS-7903464: added virtual threads executor
